Refactor location of bridge code to allow out of process bridging
Fix some issues Andras has seen
